Separated the program context functions
